@@ -2,27 +2,27 @@
 
 namespace Database\Factories;
 
-use App\Models\Todo;
+use App\Models\User;
+use App\Models\Category;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Todo>
+ */
 class TodoFactory extends Factory
 {
     /**
-     * The name of the factory's corresponding model.
-     *
-     * @var string
-     */
-    protected $model = Todo::class;
-
-    /**
      * Define the model's default state.
      *
-     * @return array
+     * @return array<string, mixed>
      */
-    public function definition() {
+    public function definition(): array
+    {
         return [
-            'item' => $this->faker->sentence(),
-            'completed' => 0
+            'user_id' => User::factory(),
+            'todo' => fake()->paragraph(),
+            'is_completed' => fake()->boolean(),
+            'category_id' => Category::factory(),
         ];
     }
 }
