@@ -3,18 +3,21 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
+use Database\Seeders\TodoSeeder;
+use Database\Seeders\CategorySeeder;
+use Database\Seeders\UserSeeder;
 
 class DatabaseSeeder extends Seeder
 {
     /**
      * Seed the application's database.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
-        // \App\Models\User::factory(10)->create();
-        $this->call(UsersTableSeeder::class);
-        $this->call(TodosTableSeeder::class);
+        $this->call([
+            UserSeeder::class,
+            CategorySeeder::class,
+            TodoSeeder::class,
+        ]);
     }
 }
