@@ -4,41 +4,53 @@ namespace Tests\Feature\Auth;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Livewire\Volt\Volt;
 use Tests\TestCase;
 
 class PasswordConfirmationTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_confirm_password_screen_can_be_rendered()
+    public function test_confirm_password_screen_can_be_rendered(): void
     {
         $user = User::factory()->create();
 
         $response = $this->actingAs($user)->get('/confirm-password');
 
-        $response->assertStatus(200);
+        $response
+            ->assertSeeVolt('pages.auth.confirm-password')
+            ->assertStatus(200);
     }
 
-    public function test_password_can_be_confirmed()
+    public function test_password_can_be_confirmed(): void
     {
         $user = User::factory()->create();
 
-        $response = $this->actingAs($user)->post('/confirm-password', [
-            'password' => 'password',
-        ]);
+        $this->actingAs($user);
 
-        $response->assertRedirect();
-        $response->assertSessionHasNoErrors();
+        $component = Volt::test('pages.auth.confirm-password')
+            ->set('password', 'password');
+
+        $component->call('confirmPassword');
+
+        $component
+            ->assertRedirect('/dashboard')
+            ->assertHasNoErrors();
     }
 
-    public function test_password_is_not_confirmed_with_invalid_password()
+    public function test_password_is_not_confirmed_with_invalid_password(): void
     {
         $user = User::factory()->create();
 
-        $response = $this->actingAs($user)->post('/confirm-password', [
-            'password' => 'wrong-password',
-        ]);
+        $this->actingAs($user);
 
-        $response->assertSessionHasErrors();
+        $component = Volt::test('pages.auth.confirm-password')
+            ->set('password', 'wrong-password');
+
+        $component->call('confirmPassword');
+
+        $component
+            ->assertNoRedirect()
+            ->assertHasErrors('password');
     }
 }
